services/1-dataset-download/requirements.txt
kagglehub
google-cloud-storage
